refactor: update identity module references to use the new model module
