+++ src/cli/switch.rs
@@ -3,7 +3,7 @@
 use clap::Args;
 
 use crate::error::AppError;
-use crate::identity::identity;
+use crate::identity::model;
 
 #[derive(Args)]
 pub struct SwitchArgs {
@@ -12,7 +12,7 @@ pub struct SwitchArgs {
 }
 
 pub fn run(args: SwitchArgs) -> Result<(), AppError> {
-    let identity = identity::resolve_identity_scope(&args.identity).ok_or_else(|| {
+    let identity = model::resolve_identity_scope(&args.identity).ok_or_else(|| {
         AppError::InvalidIdentityScope(format!(
             "invalid identity scope '{}'. Valid: personal (p), work (w)",
             args.identity
