+++ src/identity/mod.rs
@@ -1,5 +1,5 @@
 pub mod file_store;
 pub mod git_cli;
 pub mod git_config;
-pub mod identity;
+pub mod model;
 pub mod store;
